refactor : Added static methods in calculator and required test cases for intermediate calculator
