+++ tests/test_calculator.py
@@ -1,18 +1,18 @@
-''' My Calculator tests'''
-from calculator import add, subtract, multiply, divide
+'''My Calculator Test'''
+from calculator import Calculator
 
 def test_addition():
     '''Test that addition function works '''    
-    assert add(2,2) == 4
+    assert Calculator.add(2,2) == 4
 
 def test_subtraction():
-    '''Test that addition function works '''    
-    assert subtract(2,2) == 0
+    '''Test that subtraction function works '''    
+    assert Calculator.subtract(2,2) == 0
+
+def test_multiply():
+    '''Test that multiplication function works '''    
+    assert Calculator.multiply(2,2) == 4
 
 def test_divide():
     '''Test that division function works '''    
-    assert divide(2,2) == 1
-
-def test_multiply():
-    '''Test that multiply function works '''    
-    assert multiply(2,2) == 4
+    assert Calculator.divide(2,2) == 1
