+++ tests/test_operations.py
@@ -1,0 +1,18 @@
+'''My Calculator Test'''
+from calculator.operations import add, multiply, subtract, divide
+
+def test_addition():
+    '''Test that addition function works '''    
+    assert add(2,2) == 4
+
+def test_subtraction():
+    '''Test that subtraction function works '''    
+    assert subtract(2,2) == 0
+
+def test_multiplication():
+    '''Test that multiplication function works '''
+    assert multiply(2,2) == 4
+
+def test_division():
+    '''Test that division function works '''
+    assert divide(2,2) == 1
